backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
Pillow==11.0.0
psycopg2-binary==2.9.10
django-filter==24.3
django-rest-knox==5.0.2
cryptography==44.0.0
celery==5.4.0
redis==5.2.1
stripe==11.2.0
drf-spectacular==0.28.0
